♻️ refactor: type summary
