+++ .prettierignore
@@ -1,0 +1,1 @@
+**/di.config.ts
