+++ src/app/middlewares/verifyAccessToken.ts
@@ -3,7 +3,7 @@ import jwt from "jsonwebtoken";
 
 import Config from "../../config";
 
-import type { User } from "../../types/database";
+import type { User } from "../../types/user";
 import type { RequestHandler } from "express";
 
 export const verifyAccessToken: RequestHandler = (req, res, next) => {
