+++ src/types/contents.ts
@@ -1,0 +1,3 @@
+export type Contents = {
+  [key: string]: string;
+};
