+++ src/types/global.ts
@@ -1,18 +1,9 @@
-import { User } from "./collection";
+import { User } from "./user";
 
 declare global {
   namespace Express {
     export interface Request {
       user: User;
-    }
-  }
-
-  namespace NodeJS {
-    interface ProcessEnv {
-      MONGODB_URI: string;
-      JWT_SECRET_KEY: string;
-      GITHUB_CLIENT_ID: string;
-      GITHUB_CLIENT_SECRET: string;
     }
   }
 }
