+++ src/types/user.ts
@@ -1,0 +1,8 @@
+export type User = {
+  _id?: string;
+  username: string;
+  userGithubUri: string;
+  userImage?: string;
+  githubAccessToken: string;
+  projects?: string[];
+};
